Bluetooth actually works. Still needs refactoring.
